requirements.txt
fastapi==0.115.0
httpx==0.27.2
google-generativeai==0.7.2
pydantic==2.9.2
python-dotenv==1.0.1
uvicorn==0.30.6
